worked on the web3 installation
