+++ package.json
@@ -21,7 +21,8 @@
     "react-icons": "^5.2.1",
     "react-router-dom": "^6.26.0",
     "tailwind-merge": "^2.4.0",
-    "tailwindcss-animate": "^1.0.7"
+    "tailwindcss-animate": "^1.0.7",
+    "web3": "^4.11.1"
   },
   "devDependencies": {
     "@types/node": "^22.1.0",
