+++ src/pages/index.js
@@ -4,14 +4,16 @@ import Resources from "./Resources";
 import Rooms from "./Rooms";
 import Therapist from "./Therapist";
 import Communities  from "./Communities";
+import Authentication from "./Authentication";
 
 
 
 export {
-    Dashboard,
-    JoinCommunities,
-    Resources,
-    Rooms,
-    Therapist,
-    Communities
-}
+  Dashboard,
+  JoinCommunities,
+  Resources,
+  Rooms,
+  Therapist,
+  Communities,
+  Authentication
+};
